adminpanel/pyproject.toml
[tool.poetry]
name = "yap-graduate-work"
version = "0.1.0"
description = ""
authors = ["mishastik78 <69627188+mishastik78@users.noreply.github.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.5"
requests = "^2.28.2"
gunicorn = "^20.1.0"
uvicorn = "^0.20.0"
kafka-python = "^2.0.2"


[tool.poetry.group.dev.dependencies]
psycopg2-binary = "^2.9.5"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"
